+++ django_fido/__init__.py
@@ -1,2 +1,2 @@
 """Django application for FIDO protocol."""
-__version__ = '0.37'
+__version__ = '0.38'